Moved MakeCSV to InterfaceAdapter package
